Rename existing SSH implementation
to be libssh-specific
