+++ kermit/k95/ckolssh.h
