+++ kermit/k95/ckolsshs.h
